RSS: fix visibility of 'split content' content in RSS feeds.
The implementation to reveal content is carried out client-side, and we can't
do that in a RSS feed. 'Split content' and 'members only' content does not
appear in the RSS feed, and hopefully this will cause people to go to the
actual website and view the content in situ (with Coil!).

For CWP-64