Fix donation bar not showing on new installs.
The Customizer's default value is true, which causes the control to render
appropriately. When the option is fetched on a new install before any
settings changes are made, it does not exist in the database, so defaults
to false.

This leads to some user confusion between the real default behaviour.

Fixes #67